requirements.txt
kafka-python
simplejson
paramiko
openstacksdk
